LayoutTests/inspector/console/console-external-array-expected.txt

CONSOLE MESSAGE: line 9: [object Int8Array]
CONSOLE MESSAGE: line 10: [object Int16Array]
CONSOLE MESSAGE: line 11: [object Int32Array]
CONSOLE MESSAGE: line 12: [object Float32Array]
CONSOLE MESSAGE: line 13: [object Float64Array]
CONSOLE MESSAGE: line 15: [object Int8Array]
CONSOLE MESSAGE: line 16: [object Int16Array]
CONSOLE MESSAGE: line 17: [object Int32Array]
CONSOLE MESSAGE: line 18: [object Float32Array]
CONSOLE MESSAGE: line 19: [object Float64Array]
Tests that console logging detects external arrays as arrays.

[0, 0, 0, 0, 0, 0, 0, 0, 0, 0] console-external-array.html:9
[0, 0, 0, 0, 0, 0, 0, 0, 0, 0] console-external-array.html:10
[0, 0, 0, 0, 0, 0, 0, 0, 0, 0] console-external-array.html:11
[0, 0, 0, 0, 0, 0, 0, 0, 0, 0] console-external-array.html:12
[0, 0, 0, 0, 0, 0, 0, 0, 0, 0] console-external-array.html:13
Int8Array[10] console-external-array.html:15
Int16Array[10] console-external-array.html:16
Int32Array[10] console-external-array.html:17
Float32Array[10] console-external-array.html:18
Float64Array[10] console-external-array.html:19
